Updated fillCache, moved test to new class
