+++ integrationtesting/opendcs-tests/src/test/java/org/opendcs/regression_tests/CompDependsDaoTestIT.java
@@ -3,9 +3,7 @@ package org.opendcs.regression_tests;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.io.File;
-import java.util.List;
 import java.util.concurrent.TimeUnit;
-import java.util.function.Predicate;
 
 import org.junit.jupiter.api.Test;
 import org.opendcs.fixtures.AppTestBase;
@@ -91,66 +89,6 @@ final class CompDependsDaoTestIT extends AppTestBase
                         "Expected values were not found in the comp depends table."
                 );
             }
-        }
-    }
-
-    @Test
-    void testComputationFilter() throws Exception
-    {
-        try (TimeSeriesDAI tsDAI = db.makeTimeSeriesDAO();
-            ComputationDAI compDAI = db.makeComputationDAO())
-        {
-            TimeSeriesIdentifier tsIdInput = db.makeEmptyTsId();
-            TimeSeriesIdentifier tsIdOutput = db.makeEmptyTsId();
-            tsIdInput.setUniqueString("TESTSITE1.Stage.Inst.15Minutes.0.raw");
-            tsIdOutput.setUniqueString("TESTSITE1.Flow.Inst.15Minutes.0.raw");
-            DbKey inputKey = tsDAI.createTimeSeries(tsIdInput);
-            DbKey outputKey = tsDAI.createTimeSeries(tsIdOutput);
-
-            DbComputation comp = new DbComputation(DbKey.NullKey, "stage_flow_tmp");
-            DbCompParm input = new DbCompParm("indep", inputKey, tsIdInput.getInterval(), tsIdInput.getTableSelector(), 0);
-            DbCompParm output = new DbCompParm("dep", outputKey, tsIdOutput.getInterval(), tsIdOutput.getTableSelector(), 0);
-            comp.addParm(input);
-            comp.addParm(output);
-            comp.setAlgorithmName("TabRating");
-            comp.setApplicationName("compproc_regtest");
-            comp.setEnabled(true);
-
-            compDAI.writeComputation(comp);
-            final DbComputation compInDb = compDAI.getComputationByName("stage_flow_tmp");
-
-            // retrieve with default filter
-            List<DbComputation> comps = compDAI.listComps(item -> true);
-            assertTrue(comps.stream().anyMatch(c -> c.getKey().equals(compInDb.getKey())));
-
-            // retrieve with filter
-            Predicate<DbComputation> filter = item -> {
-              for (DbCompParm parm : item.getParmList())
-              {
-                if (parm.getRoleName().equalsIgnoreCase("indep"))
-                {
-                  return true;
-                }
-              }
-                return false;
-            };
-
-            comps = compDAI.listComps(filter);
-            assertTrue(comps.stream().anyMatch(c -> c.getKey().equals(compInDb.getKey())));
-
-            // retrieve with filter that should not match
-            filter = item -> {
-              for (DbCompParm parm : item.getParmList())
-              {
-                if (parm.getRoleName().equalsIgnoreCase("foo"))
-                {
-                  return true;
-                }
-              }
-                return false;
-            };
-            comps = compDAI.listComps(filter);
-            assertTrue(comps.stream().noneMatch(c -> c.getKey().equals(compInDb.getKey())));
         }
     }
 }
