+++ java/opendcs/src/main/java/opendcs/dao/ComputationDAO.java
@@ -172,91 +172,99 @@ public class ComputationDAO
 
 	private void fillCache()
 	{
-		debug1("ComputationDAO.fillCache()");
-
-		String q = "select " + compTableColumns + " from CP_COMPUTATION ";
-
-		try
+		if (System.currentTimeMillis() - lastCacheFill > CACHE_TIME_LIMIT)
 		{
-			// clear the cache as we're rebuilding it. Though analyzing usage, this may never
-			// be required.
-			compCache.clear();
-			final int[] n = new int[1];
-			n[0] = 0;
-			doQuery(q, rs ->
+			debug1("ComputationDAO.fillCache()");
+
+			String q = "select " + compTableColumns + " from CP_COMPUTATION ";
+
+			try
 			{
-				compCache.put(rs2comp(rs));
-				n[0]++;
-			});
-			log.debug("{} cp_computation recs read.", n[0]);
-
-			n[0] = propsDao.readPropertiesIntoCache("CP_COMP_PROPERTY", compCache);
-			log.debug("{} cp_comp_property recs read.", n[0]);
-
-			ArrayList<CompAppInfo> apps = loadingAppDAO.listComputationApps(true);
-			ArrayList<DbCompAlgorithm> algos = algorithmDAO.listAlgorithms();
-
-			// Associate comps with groups, apps & algorithms.
-			for(CacheIterator it = compCache.iterator(); it.hasNext(); )
-			{
-				DbComputation comp = (DbComputation)it.next();
-				if (!DbKey.isNull(comp.getGroupId()))
-					comp.setGroup(tsGroupDAO.getTsGroupById(comp.getGroupId()));
-
-				if (!DbKey.isNull(comp.getAppId()))
-					for(CompAppInfo cai : apps)
-						if (comp.getAppId().equals(cai.getAppId()))
-							comp.setApplicationName(cai.getAppName());
-
-				if (!DbKey.isNull(comp.getAlgorithmId()))
-					for(DbCompAlgorithm algo : algos)
-						if (comp.getAlgorithmId().equals(algo.getId()))
-						{
-							comp.setAlgorithm(algo);
-							comp.setAlgorithmName(algo.getName());
-						}
-			}
-
-			// Note the parms rely on the algorithms being in place. So get them now.
-			q = "select a.* from CP_COMP_TS_PARM a, CP_COMPUTATION b "
-				+ "where a.COMPUTATION_ID = b.COMPUTATION_ID";
-			n[0] = 0;
-			doQuery(q, rs ->
-			{
-				DbKey compId = DbKey.createDbKey(rs, 1);
-				DbComputation comp = compCache.getByKey(compId);
-				if (comp == null)
+				// clear the cache as we're rebuilding it. Though analyzing usage, this may never
+				// be required.
+				compCache.clear();
+				final int[] n = new int[1];
+				n[0] = 0;
+				doQuery(q, rs ->
 				{
-					log.warn("CP_COMP_TS_PARM with comp id={} with no matching computation.", compId);
-				}
-				else
-				{
-					rs2compParm(comp, rs);
+					compCache.put(rs2comp(rs));
 					n[0]++;
+				});
+				log.debug("{} cp_computation recs read.", n[0]);
+
+				n[0] = propsDao.readPropertiesIntoCache("CP_COMP_PROPERTY", compCache);
+				log.debug("{} cp_comp_property recs read.", n[0]);
+
+				ArrayList<CompAppInfo> apps = loadingAppDAO.listComputationApps(true);
+				ArrayList<DbCompAlgorithm> algos = algorithmDAO.listAlgorithms();
+
+				// Associate comps with groups, apps & algorithms.
+				for(CacheIterator it = compCache.iterator(); it.hasNext(); )
+				{
+					DbComputation comp = (DbComputation) it.next();
+					if(!DbKey.isNull(comp.getGroupId()))
+						comp.setGroup(tsGroupDAO.getTsGroupById(comp.getGroupId()));
+
+					if(!DbKey.isNull(comp.getAppId()))
+						for(CompAppInfo cai : apps)
+							if(comp.getAppId().equals(cai.getAppId()))
+								comp.setApplicationName(cai.getAppName());
+
+					if(!DbKey.isNull(comp.getAlgorithmId()))
+						for(DbCompAlgorithm algo : algos)
+							if(comp.getAlgorithmId().equals(algo.getId()))
+							{
+								comp.setAlgorithm(algo);
+								comp.setAlgorithmName(algo.getName());
+							}
 				}
-			});
-			log.debug("{} cp_comp_ts_parm recs read.", n[0]);
 
-			for(CacheIterator it = compCache.iterator(); it.hasNext(); )
-			{
-				DbComputation comp = (DbComputation)it.next();
+				// Note the parms rely on the algorithms being in place. So get them now.
+				q = "select a.* from CP_COMP_TS_PARM a, CP_COMPUTATION b "
+						+ "where a.COMPUTATION_ID = b.COMPUTATION_ID";
+				n[0] = 0;
+				doQuery(q, rs ->
+				{
+					DbKey compId = DbKey.createDbKey(rs, 1);
+					DbComputation comp = compCache.getByKey(compId);
+					if(comp == null)
+					{
+						log.warn("CP_COMP_TS_PARM with comp id={} with no matching computation.", compId);
+					}
+					else
+					{
+						rs2compParm(comp, rs);
+						n[0]++;
+					}
+				});
+				log.debug("{} cp_comp_ts_parm recs read.", n[0]);
 
-				// Make sure site IDs and datatype IDs are set in the parms
-				for(DbCompParm parm : comp.getParmList())
-					if (!parm.getSiteDataTypeId().isNull())
-						try { db.expandSDI(parm); }
-						catch (NoSuchObjectException e) {}
+				for(CacheIterator it = compCache.iterator(); it.hasNext(); )
+				{
+					DbComputation comp = (DbComputation) it.next();
+
+					// Make sure site IDs and datatype IDs are set in the parms
+					for(DbCompParm parm : comp.getParmList())
+						if(!parm.getSiteDataTypeId().isNull())
+							try
+							{
+								db.expandSDI(parm);
+							}
+							catch(NoSuchObjectException e)
+							{
+							}
+				}
+				log.debug("fillCache finished, {} computations cached.", compCache.size());
 			}
-			log.debug("fillCache finished, {} computations cached.", compCache.size());
-		}
-		catch(Exception ex)
-		{
-			log.atWarn()
-			   .setCause(ex)
-			   .log("Exception filling computation hash.");
-		}
+			catch(Exception ex)
+			{
+				log.atWarn()
+						.setCause(ex)
+						.log("Exception filling computation hash.");
+			}
 
-		lastCacheFill = System.currentTimeMillis();
+			lastCacheFill = System.currentTimeMillis();
+		}
 	}
 
 	@Override
@@ -563,10 +571,7 @@ public class ComputationDAO
 	{
 		debug1("listCompsForGUI " + filter);
 
-		if (System.currentTimeMillis() - lastCacheFill > CACHE_TIME_LIMIT)
-		{
-			fillCache();
-		}
+		fillCache();
 
 		ArrayList<DbComputation> ret = new ArrayList<>();
 		for (CacheIterator it = compCache.iterator(); it.hasNext(); )
@@ -604,10 +609,7 @@ public class ComputationDAO
 	{
 		debug1("listComps " + filter);
 
-		if (System.currentTimeMillis() - lastCacheFill > CACHE_TIME_LIMIT)
-		{
-			fillCache();
-		}
+		fillCache();
 
 		List<DbComputation> ret = new ArrayList<>();
 		for (DbObjectCache<DbComputation>.CacheIterator it = compCache.iterator(); it.hasNext(); )
